functions/getMasterFile/requirements.txt
Flask==3.1.2
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==3.4.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
Flask==3.1.2
protobuf==6.32.1